First version of multiple-GPU histogram calculation.
First attempt to do a multiple-GPU version (Re. #3).
Here, we make one extra step in computing the histograms. We first brak
down image into "stripes" (multiple rows). Then for each "stripe"
we compute its own histogram (that also includes breaking down to blocks
and then accumulating the resulting histograms).
To accumulate histograms from different stripes, one device takes the
"master" role, gathers "strip" histograms from all devices and accumulates
them.
Unfortunately, that "gathering" step cannot be performed on the machine
I am working on - it responds with "peer access not available" when I
try to enable peer access.
Of course, that device-to-device copying could be modeled with the
device-host-device, but that would create a tight bottleneck. I wonder
if in that case, accumulating on CPU would be faster than copying-to-cpu,
copying-to-another-device, accumulating, copying-to-cpu again.
I'll see if I find a better idea.